cleanup DataStreamInterface for PR
